sprint mission 5

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicReadStatusService.java
@@ -1,72 +1,77 @@
 package com.sprint.mission.discodeit.service.basic;
 
-import com.sprint.mission.discodeit.dto.readStatus.ReadStatusCreateRequest;
-import com.sprint.mission.discodeit.dto.readStatus.ReadStatusDto;
-import com.sprint.mission.discodeit.dto.readStatus.ReadStatusUpdateRequest;
+import com.sprint.mission.discodeit.dto.request.ReadStatusCreateRequest;
+import com.sprint.mission.discodeit.dto.request.ReadStatusUpdateRequest;
 import com.sprint.mission.discodeit.entity.ReadStatus;
-import com.sprint.mission.discodeit.exception.notfound.ResourceNotFoundException;
-import com.sprint.mission.discodeit.mapper.ReadStatusMapper;
+import com.sprint.mission.discodeit.repository.ChannelRepository;
 import com.sprint.mission.discodeit.repository.ReadStatusRepository;
+import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.service.ReadStatusService;
-import com.sprint.mission.discodeit.validation.ReadStatusValidator;
 import java.time.Instant;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-@Service
 @RequiredArgsConstructor
+@Service
 public class BasicReadStatusService implements ReadStatusService {
 
   private final ReadStatusRepository readStatusRepository;
-  private final ReadStatusValidator readStatusValidator;
-  private final ReadStatusMapper readStatusMapper;
+  private final UserRepository userRepository;
+  private final ChannelRepository channelRepository;
 
   @Override
-  public ReadStatusDto create(ReadStatusCreateRequest request) {
-
+  public ReadStatus create(ReadStatusCreateRequest request) {
     UUID userId = request.userId();
     UUID channelId = request.channelId();
-    Instant lastReadTime = request.lastReadTime();
 
-    readStatusValidator.validateUser(userId);
-    readStatusValidator.validateChannel(channelId);
-    readStatusValidator.validateDuplicateReadStatus(userId, channelId);
+    if (!userRepository.existsById(userId)) {
+      throw new NoSuchElementException("User with id " + userId + " does not exist");
+    }
+    if (!channelRepository.existsById(channelId)) {
+      throw new NoSuchElementException("Channel with id " + channelId + " does not exist");
+    }
+    if (readStatusRepository.findAllByUserId(userId).stream()
+        .anyMatch(readStatus -> readStatus.getChannelId().equals(channelId))) {
+      throw new IllegalArgumentException(
+          "ReadStatus with userId " + userId + " and channelId " + channelId + " already exists");
+    }
 
-    ReadStatus readStatus = new ReadStatus(userId, channelId, lastReadTime);
-    ReadStatus createdReadStatus = readStatusRepository.save(readStatus);
-
-    return readStatusMapper.readStatusEntityToDto(createdReadStatus);
+    Instant lastReadAt = request.lastReadAt();
+    ReadStatus readStatus = new ReadStatus(userId, channelId, lastReadAt);
+    return readStatusRepository.save(readStatus);
   }
 
   @Override
-  public ReadStatusDto findById(UUID readStatusId) {
+  public ReadStatus find(UUID readStatusId) {
     return readStatusRepository.findById(readStatusId)
-        .map(readStatusMapper::readStatusEntityToDto)
-        .orElseThrow(() -> new ResourceNotFoundException("ReadStatus not found: " + readStatusId));
+        .orElseThrow(
+            () -> new NoSuchElementException("ReadStatus with id " + readStatusId + " not found"));
   }
 
   @Override
-  public List<ReadStatusDto> findAllByUserId(UUID userId) {
+  public List<ReadStatus> findAllByUserId(UUID userId) {
     return readStatusRepository.findAllByUserId(userId).stream()
-        .map(readStatusMapper::readStatusEntityToDto)
         .toList();
   }
 
   @Override
-  public ReadStatusDto update(UUID readStatusId, ReadStatusUpdateRequest request) {
-    Instant newLastReadTime = request.newLastReadTime();
+  public ReadStatus update(UUID readStatusId, ReadStatusUpdateRequest request) {
+    Instant newLastReadAt = request.newLastReadAt();
     ReadStatus readStatus = readStatusRepository.findById(readStatusId)
-        .orElseThrow(() -> new ResourceNotFoundException("Read status not found: " + readStatusId));
-    readStatus.update(newLastReadTime);
-    ReadStatus updatedReadStatus = readStatusRepository.save(readStatus);
-    return readStatusMapper.readStatusEntityToDto(updatedReadStatus);
+        .orElseThrow(
+            () -> new NoSuchElementException("ReadStatus with id " + readStatusId + " not found"));
+    readStatus.update(newLastReadAt);
+    return readStatusRepository.save(readStatus);
   }
 
   @Override
   public void delete(UUID readStatusId) {
-    readStatusValidator.validateReadStatus(readStatusId);
+    if (!readStatusRepository.existsById(readStatusId)) {
+      throw new NoSuchElementException("ReadStatus with id " + readStatusId + " not found");
+    }
     readStatusRepository.deleteById(readStatusId);
   }
 }
